refact: extracted update parser login
